requirements.txt
langchain
openai
chromadb
tiktoken
tabulate
sqlalchemy
sqlalchemy-bigquery
google-cloud-bigquery
flask
gunicorn
BeautifulSoup4
colorama